Retrieve forwardable kerb tkt in thread safe way
This changes how a forwardable Kerberos ticket is retrieved by using the
Kerberos API directly instead of using GSSAPI. The original method
relied on environment variables which are not thread safe. By using the
Kerberos API this can retrieve the credentials in any way desired and
store them in a thread safe context.