Add type annotations and context extra info (#30)
Finalises type annotations in the library and exports the py.typed stub
for external libraries to utilise. Also adds mypy to CI checks to
validate the type annotations that are currently in place.

Also adds `get_extra_info` to the context proxy to allow the caller to
retrieve context specific information. Currently this is used by CredSSP
to expose the delegated credentials from the initiator and both the SSL
context and object used during the handshake.