CredSSP min protocol and stage information
Added the ability to specify the minimum CredSSP protocol version that
the caller allows. Talking to a peer that does not meet this minimum
requirement will result in an InvalidTokenError. This allows the caller
to only allow peers that have mitigated CVE-2018-0886 or any future CVEs
that result in a new protocol version.

Also adds the auth_stage and protocol_version extra info for a CredSSP
context proxy allowing the caller to inspect the protocol version that
was ultimately negotiated between the initiator and acceptor as well as
provide more details on the authentication process in CredSSP for more
context around error messages.