Fix Heap Allocation Errors on Windows
Fix errors when running with heap allocation checks on Windows. This
moves all the management of the SecBuffer objects to the Python code to
keep the Cython management simple. It also exposed some of the
inefficiencies that are in place where memory was copied to avoid it
mutating which can be addressed in a separate PR.