Add new_context() option to copy a context
Adds the new_context() function that will create a new context based on
the existing contexts credentials and options. This is designed to be a
more efficient way to recreate a brand new security context. For example
it can recreate the context with a cached provider specific credential
rather than retrieving a new one for each context. This is important
when using Kerberos authentication as the retrieved TGT can be cached
and reused across multiple contexts.