Migrate SSPI to external library
This commit adds a new dependency for Windows that includes the sspi
bindings. This removes the extra Cython build step for Windows allowing
it to be shipped as a pure wheel and simplifying our build steps. It
also solves the problem with Cython 3 that our original Cython code was
incompatible with. Finally it removes extra code allowing it to be
tested in a more self contained library.